[Cleanup] Unify vllm.utils.flashinfer and flashinfer_utils
Move pure utility functions from flashinfer_utils.py to the canonical
vllm.utils.flashinfer module. Keep MoE-layer-coupled helpers in
fused_moe/flashinfer_utils.py as a backward-compatible stub.
